Split helper fn by different files
