+++ helpers/index.js
@@ -1,143 +1,23 @@
 'use strict'
 
-const fs = require('fs')
-const { Grape } = require('grenache-grape')
-const waterfall = require('async/waterfall')
-const fp = require('find-free-port')
-const DHT = require('bittorrent-dht')
-
-const ports = {
-  grape1DhtPort: 20002,
-  grape1ApiPort: 40001,
-  grape2DhtPort: 20001,
-  grape2ApiPort: 30001,
-  workerApiPort: 1337,
-  workerWsPort: 1455,
-  expressApiPort: 34343
-}
-
-const getDefaultPorts = () => {
-  return { ...ports }
-}
-
-const _asyncForEach = async (array, callback) => {
-  for (let index = 0; index < array.length; index++) {
-    await callback(array[index], index, array)
-  }
-}
-
-const _checkPortsUniq = (port, ports = {}) => {
-  return Object.entries(ports).every(([key, nextPort]) => port !== nextPort)
-}
-
-const getFreePort = async (ports = {}) => {
-  const res = {}
-
-  await _asyncForEach(Object.entries(ports), async ([key, port]) => {
-    let count = 0
-
-    while (true) {
-      count += 1
-
-      if (count > 100) throw new Error('NO_FREE_PORT')
-
-      const freePort = (await fp(port, '127.0.0.1'))[0]
-      if (/dht/i.test(key)) {
-        const dht = new DHT()
-
-        try {
-          await new Promise((resolve, reject) => {
-            dht.once('error', (err) => {
-              dht.removeListener('listening', resolve)
-              reject(err)
-            })
-            dht.once('listening', () => {
-              dht.removeListener('error', reject)
-              resolve()
-            })
-            dht.listen(port)
-          })
-
-          dht.destroy()
-        } catch (err) {
-          port += 1
-          dht.destroy()
-
-          continue
-        }
-      }
-
-      if (_checkPortsUniq(freePort, res)) {
-        res[key] = freePort
-
-        break
-      }
-
-      port += 1
-    }
-  })
-
-  return res
-}
-
-const checkAndChangeAccess = (path) => {
-  try {
-    fs.accessSync(path, fs.constants.R_OK | fs.constants.W_OK)
-  } catch (err) {
-    fs.chmodSync(path, '766')
-  }
-}
-
-const bootTwoGrapes = (locPorts = ports) => {
-  return new Promise((resolve, reject) => {
-    const confGrape1 = {
-      dht_port: locPorts.grape1DhtPort,
-      dht_bootstrap: [`127.0.0.1:${locPorts.grape2DhtPort}`],
-      api_port: locPorts.grape1ApiPort
-    }
-    const confGrape2 = {
-      dht_port: locPorts.grape2DhtPort,
-      dht_bootstrap: [`127.0.0.1:${locPorts.grape1DhtPort}`],
-      api_port: locPorts.grape2ApiPort
-    }
-
-    const grape1 = new Grape(confGrape1)
-    const grape2 = new Grape(confGrape2)
-
-    waterfall(
-      [
-        cb => {
-          grape1.start()
-          grape1.once('ready', cb)
-        },
-        cb => {
-          grape2.start()
-          grape2.once('node', cb)
-        }
-      ],
-      (err) => {
-        if (err) reject(err)
-
-        resolve([grape1, grape2])
-      }
-    )
-  })
-}
-
-const killGrapes = (grapes, done = () => {}) => {
-  grapes[0].stop(err => {
-    if (err) throw err
-    grapes[1].stop(err => {
-      if (err) throw err
-      done()
-    })
-  })
-}
+const {
+  bootTwoGrapes,
+  killGrapes
+} = require('./grapes')
+const {
+  getFreePort,
+  getDefaultPorts
+} = require('./ports')
+const {
+  checkAndChangeAccess
+} = require('./utils')
+const windowStateKeeper = require('./window-state-keeper')
 
 module.exports = {
   bootTwoGrapes,
   killGrapes,
   getFreePort,
   getDefaultPorts,
-  checkAndChangeAccess
+  checkAndChangeAccess,
+  windowStateKeeper
 }
