+++ helpers/window-state-keeper.js
